+++ demisto_sdk/commands/validate/validators/PB_validators/PB118_is_input_key_not_in_tasks.py
@@ -4,13 +4,13 @@ from typing import Iterable, List
 
 from demisto_sdk.commands.common.constants import GitStatuses
 from demisto_sdk.commands.content_graph.objects.playbook import Playbook
+from demisto_sdk.commands.validate.tools import (
+    collect_all_inputs_from_inputs_section,
+    collect_all_inputs_in_use,
+)
 from demisto_sdk.commands.validate.validators.base_validator import (
     BaseValidator,
     ValidationResult,
-)
-from demisto_sdk.commands.validate.validators.tools import (
-    collect_all_inputs_from_inputs_section,
-    collect_all_inputs_in_use,
 )
 
 ContentTypes = Playbook
